refactor: move MusicBand to Long id

+++ src/main/java/org/is/bandmanager/repository/BestBandAwardRepository.java
@@ -21,7 +21,7 @@ public interface BestBandAwardRepository extends JpaRepository<BestBandAward, Lo
         return findAll(specification, pageable);
     }
 
-    List<BestBandAward> deleteAllByBandIdIn(List<Integer> bandIds);
+    List<BestBandAward> deleteAllByBandIdIn(List<Long> bandIds);
 
 }
